+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # probeye changelog
 
-## 2.2.0 (2022-Apr-XX)
+## 2.2.0 (2022-Apr-26)
 ### Changed
 - The forward model's 'definition'-method was renamed to 'interface'.
 - The user-definition of the forward model's 'interface'-method is now mandatory.
@@ -10,6 +10,7 @@
 - Removed method InverseProblem.transform_experimental_data since it is no longer necessary without the pyro/torch-solver.
 - The specification of a parameter's type (model, prior or likelihood) is no longer necessary (it can be detected automatically).
 - The multivariate normal prior is now an own prior type and not anymore included in the normal prior class.
+- Updated the docs (however, still work to be done).
 
 ## 2.1.5 (2022-Mar-31)
 ### Changed